Update README.md
README page with the following info:
a. Instructions on how to build the project.
b. Adapted design pattern for the app architecture with justification on why to
choose it.
c. Adapted image loader library with justification on why to choose it.
d. Used database in the app with justification on why to choose it.